changed how card "flip"
Instead of making two overlapping gameobject with different sprites, I save both sprites to variables on the card and switch between them. For some reason it will not show the second card you click until after the comparison is made and it "flips" back. This shouldn't be possible as for it to make the comparison it has to run the code that changes the card sprite. I'm guessing it has something to do with not running certain types of code while executing a loop but I really have no idea. I probably need to shelf this and make a new game that isn't all generated from a script like an idiot.